requirements.txt
absl-py==0.3.0
astor==0.7.1
azure-functions==1.0.0b4
azure-functions-worker==1.0.0b4
certifi==2018.4.16
joblib==0.13.2
chardet==3.0.4
distlib==0.2.7
gast==0.2.0
grpcio==1.12.1
grpcio-tools==1.12.1
h5py==2.8.0
idna==2.7
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
Markdown==2.6.11
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.0
PyYAML==3.12
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
urllib3==1.23
Werkzeug==0.14.1
